fix: @types directory moved to root
