+++ @types/Project.d.ts
@@ -1,4 +1,4 @@
-import { RepositoryFragment } from "../schema/github.graphql";
+import { RepositoryFragment } from "../app/schema/github.graphql";
 
 interface Project extends RepositoryFragment {
   languages?: (string | undefined)[];
