+++ app/actions/getPostsByTag.ts
@@ -4,7 +4,7 @@ import {
   TagPostsByPublicationDocument,
   TagPostsByPublicationQuery,
 } from "@/app/schema/hashnode.graphql";
-import PostsInfo from "../@types/PostsInfo";
+import PostsInfo from "../../@types/PostsInfo";
 
 async function getPostsByTag(
   tagSlug: string,
