+++ app/components/BlogCardContainer.tsx
@@ -4,7 +4,7 @@ import { useCallback, useState } from "react";
 import { PageInfo, PostFragment } from "@/app/schema/hashnode.graphql";
 import BlogCard from "@/app/components/BlogCard";
 import CardContainer from "@/app/components/CardContainer";
-import PostsInfo from "../@types/PostsInfo";
+import PostsInfo from "../../@types/PostsInfo";
 import getPostsByTag from "../actions/getPostsByTag";
 import getPostsBySeries from "../actions/getPostsBySeries";
 import getAllPosts from "../actions/getAllPosts";
